feat(scripts): move generators related modules under generators/
