+++ package.json
@@ -19,7 +19,7 @@
     "buildto": "lage build --verbose --to",
     "buildto:lerna": "node ./scripts/monorepo/buildTo.js",
     "bundle": "lage bundle --verbose",
-    "bundlesizecollect": "node ./scripts/bundle-size-collect",
+    "bundlesizecollect": "node ./scripts/generators/bundle-size-collect",
     "change": "beachball change --no-commit",
     "check:change": "beachball check",
     "check:modified-files": "yarn workspace @fluentui/scripts just check-for-modified-files",
@@ -27,12 +27,12 @@
     "check:installed-dependencies-versions": "satisfied --skip-invalid --ignore \"prettier|angular|lit|sass|@storybook/html|@storybook/mdx2-csf|svelte|@testing-library|vue|@cypress/react|cypress|@swc/wasm|@cactuslab/usepubsub\"",
     "clean": "lage clean --verbose",
     "code-style": "lage code-style --verbose",
-    "codepen": "cd packages/react && node ../../scripts/local-codepen.js",
-    "copy-notices": "node scripts/copy-notices.js",
+    "codepen": "cd packages/react && node ../../scripts/executors/local-codepen.js",
+    "copy-notices": "node scripts/generators/copy-notices.js",
     "create-component": "plop --plopfile ./scripts/generators/create-component/index.ts --dest . --require ./scripts/ts-node-register",
     "create-package": "plop --plopfile ./scripts/generators/create-package/index.ts --dest . --require ./scripts/ts-node-register",
     "e2e": "lage e2e --verbose --concurrency=1",
-    "format": "node scripts/format.js",
+    "format": "node scripts/executors/format.js",
     "generate-version-files": "yarn workspace @fluentui/scripts just generate-version-files",
     "graph": "node ./scripts/dependency-graph-generator/index.js",
     "lint": "lage lint --verbose",
@@ -43,17 +43,17 @@
     "postinstall": "node ./scripts/package-manager/postinstall.js",
     "preinstall": "node ./scripts/package-manager/use-yarn-please.js",
     "publish:beachball": "beachball publish -b origin/master --access public -y",
-    "rebuild": "node ./scripts/invalidate-just-cache.js && yarn build --reset-cache",
+    "rebuild": "node ./scripts/executors/invalidate-just-cache.js && yarn build --reset-cache",
     "release:fluentui:canary": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-canary",
     "release:fluentui:minor": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-minor",
     "release:fluentui:pack-nightly": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish pack-nightly",
     "release:fluentui:patch": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish publish-patch",
     "release:fluentui:post-validation": "node -r ./scripts/ts-node-register ./scripts/fluentui-publish post-publish",
-    "rename-package": "node -r ./scripts/ts-node-register ./scripts/rename-package.ts",
+    "rename-package": "node -r ./scripts/ts-node-register ./scripts/generators/rename-package.ts",
     "run:published": "node ./scripts/monorepo/runPublished.js",
     "runto:lerna": "node ./scripts/monorepo/runTo.js",
-    "scrub": "node ./scripts/scrub.js",
-    "start": "node scripts/start.js",
+    "scrub": "node ./scripts/executors/scrub.js",
+    "start": "node ./scripts/executors/start.js",
     "start:northstar": "yarn workspace @fluentui/docs start",
     "stats:build": "gulp stats",
     "stats:save": "gulp stats:save",
