+++ packages/react-components/react-theme/package.json
@@ -21,7 +21,7 @@
     "start": "yarn storybook",
     "storybook": "start-storybook",
     "test": "jest --passWithNoTests",
-    "token-pipeline": "node -r ../../../scripts/ts-node-register ../../../scripts/token-pipeline.ts",
+    "token-pipeline": "node -r ../../../scripts/ts-node-register ../../../scripts/generators/token-pipeline.ts",
     "type-check": "tsc -b tsconfig.json",
     "generate-api": "tsc -p ./tsconfig.lib.json --emitDeclarationOnly && just-scripts api-extractor"
   },
