+++ scripts/generators/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "types": ["node"]
   },
-  "include": ["**/*.ts"],
+  "include": ["**/*.ts", "**/*.js"],
   "exclude": ["**/plop-templates/**", "**/plop-templates-node/**", "**/plop-templates-react/**"]
 }
